web/download: revert promoting mingw-builds.
This reverts two commits:
- r6292: web/download: move mingw-builds as the first toolchain.
- r6294: web/download: note that for windows hosts, mingw-builds should
         be preferred

I've decided to mostly stop touching the download page.
I had done these two changes after having asked Kaï about it and with
the idea that there would be more changes to the corresponding section
on the download page.

I have no idea why (and probably don't want to know) but it's clear that
this section of the page doesn't change without me. And with much more
work and frustration than it should.

I do not feel comfortable making more visible something that will not be
updated.

Revert this commit if the situation changes or you believe I'm under the
wrong impression.

git-svn-id: svn+ssh://svn.code.sf.net/p/mingw-w64/code/web@6336 4407c894-4637-0410-b4f5-ada5f102cad1