web: update source link to CRTv3.
It has been some time but the download page hadn't been updated to point to
the new CRT.

Also update the page to mention that winpthreads is in CRTv3 and that it
must be fetched from SVN for previous versions.

Since winpthreads is in the main download, remove the link to the
#source-winpthreads from the navigation bar on the left of the website.

Also sneak in a small formatting fix for yypkg.

git-svn-id: svn+ssh://svn.code.sf.net/p/mingw-w64/code/web@6342 4407c894-4637-0410-b4f5-ada5f102cad1